Updated installation instructions for WSL.
The old guide is out of date and installing WSL has become easier and can now be done in a single step.  For the ease of maintainance, I've replaced the old guide by the already existing link to the official installation instructions, and added a couple of hints on how to get started.